Change SnippetGenerator class, overhaul
Add Map<String,List<Integer>> wordHashMap is a Map with the words as key and a list for positions. For searching words.
Add Map<Integer,String> positionHashMap is Map with the positions of the words as key and the word itself as value. For snippet generation.

Add Constructor SnippetGenerator(Stream<String> lines)

Add Method List<String> generateSnippets(String searchTerm, int snippetSize)